Move D3D9-specific vertex shader code out of XbVertexShader
